+++ src/main/java/nextstep/subway/line/service/LineProvider.java
@@ -1,0 +1,9 @@
+package nextstep.subway.line.service;
+
+import nextstep.subway.line.domain.Line;
+
+import java.util.List;
+
+public interface LineProvider {
+    List<Line> getAllLines();
+}
